+++ table-api/table_item_request_builder.go
@@ -1,17 +1,17 @@
 package tableapi
 
-import "github.com/michaeldcanady/servicenow-sdk-go/core"
+import (
+	"context"
 
-type TableItemRequestBuilder struct {
-	core.RequestBuilder
-}
+	"github.com/michaeldcanady/servicenow-sdk-go/core"
+)
 
-// NewTableItemRequestBuilder creates a new instance of the TableItemRequestBuilder associated with the given URL and Client.
+// NewTableItemRequestBuilder2 creates a new instance of the TableItemRequestBuilder associated with the given URL and Client.
 // It accepts the URL and Client as parameters and returns a pointer to the created TableItemRequestBuilder.
-func NewTableItemRequestBuilder(client core.Client, pathParameters map[string]string) *TableItemRequestBuilder {
-	requestBuilder := core.NewRequestBuilder(
+func NewTableItemRequestBuilder2(client core.Client2, pathParameters map[string]string) *TableItemRequestBuilder {
+	requestBuilder := core.NewRequestBuilder2(
 		client,
-		"{+baseurl}/table{/table}{/sysId}{?sysparm_display_value,sysparm_exclude_reference_link,sysparm_fields,sysparm_input_display_value,sysparm_query_no_domain,sysparm_view,sysparm_query_no_domain}",
+		tableItemURLTemplate,
 		pathParameters,
 	)
 	return &TableItemRequestBuilder{
@@ -19,15 +19,8 @@ func NewTableItemRequestBuilder(client core.Client, pathParameters map[string]st
 	}
 }
 
-// Get sends an HTTP GET request using the specified query parameters and returns a TableItemResponse.
-//
-// Parameters:
-//   - params: An instance of TableItemRequestBuilderGetQueryParameters to include in the GET request.
-//
-// Returns:
-//   - *TableItemResponse: The response data as a TableItemResponse.
-//   - error: An error if there was an issue with the request or response.
-func (rB *TableItemRequestBuilder) Get(params *TableItemRequestBuilderGetQueryParameters) (*TableItemResponse, error) {
+// Get2 sends an HTTP GET request using the specified query parameters and returns a TableItemResponse.
+func (rB *TableItemRequestBuilder) Get2(ctx context.Context, params *TableItemRequestBuilderGetQueryParameters) (*TableItemResponse, error) {
 	config := &tableItemGetRequestConfiguration2[TableEntry]{
 		header:   nil,
 		query:    params,
@@ -35,7 +28,7 @@ func (rB *TableItemRequestBuilder) Get(params *TableItemRequestBuilderGetQueryPa
 		response: &TableItemResponse2[TableEntry]{},
 	}
 
-	err := rB.SendGet2(config.toConfiguration()) //nolint:staticcheck
+	err := rB.SendGet3(ctx, config.toConfiguration())
 	if err != nil {
 		return nil, err
 	}
@@ -43,14 +36,8 @@ func (rB *TableItemRequestBuilder) Get(params *TableItemRequestBuilderGetQueryPa
 	return config.response, nil
 }
 
-// Delete sends an HTTP DELETE request using the specified query parameters and returns an error if the request or response encounters any issues.
-//
-// Parameters:
-//   - params: An instance of TableItemRequestBuilderDeleteQueryParameters to include in the DELETE request.
-//
-// Returns:
-//   - error: An error if there was an issue with the request or response, or nil if the request was successful.
-func (rB *TableItemRequestBuilder) Delete(params *TableItemRequestBuilderDeleteQueryParameters) error {
+// Delete2 sends an HTTP DELETE request using the specified query parameters and returns an error if the request or response encounters any issues.
+func (rB *TableItemRequestBuilder) Delete2(ctx context.Context, params *TableItemRequestBuilderDeleteQueryParameters) error {
 	config := &tableItemDeleteRequestConfiguration2[TableEntry]{
 		header:   nil,
 		query:    params,
@@ -58,38 +45,13 @@ func (rB *TableItemRequestBuilder) Delete(params *TableItemRequestBuilderDeleteQ
 		response: nil,
 	}
 
-	return rB.SendDelete2(config.toConfiguration())
+	return rB.SendDelete3(ctx, config.toConfiguration())
 }
 
-// Deprecated: deprecated since v1.4.0. Use `Put2` instead.
-//
-// Put updates a table item using an HTTP PUT request.
+// Put3 updates a table item using an HTTP PUT request.
 // It takes a map of table entry data and optional query parameters to send in the request.
 // The method returns a TableItemResponse representing the updated item or an error if the request fails.
-//
-// Parameters:
-//   - tableEntry: A map containing the data to update the table item.
-//   - params: An optional pointer to TableItemRequestBuilderPutQueryParameters, which can be used to specify query parameters for the request.
-//
-// Returns:
-//   - *TableItemResponse: A TableItemResponse containing the updated item data.
-//   - error: An error, if the request fails at any point, such as request information creation or JSON deserialization.
-func (rB *TableItemRequestBuilder) Put(tableEntry map[string]string, params *TableItemRequestBuilderPutQueryParameters) (*TableItemResponse, error) {
-	return rB.Put2(tableEntry, params)
-}
-
-// Put2 updates a table item using an HTTP PUT request.
-// It takes a map of table entry data and optional query parameters to send in the request.
-// The method returns a TableItemResponse representing the updated item or an error if the request fails.
-//
-// Parameters:
-//   - tableEntry: A map[string]string or TableEntry containing the data to update the table item.
-//   - params: An optional pointer to TableItemRequestBuilderPutQueryParameters, which can be used to specify query parameters for the request.
-//
-// Returns:
-//   - *TableItemResponse: A TableItemResponse containing the updated item data.
-//   - error: An error, if the request fails at any point, such as request information creation or JSON deserialization.
-func (rB *TableItemRequestBuilder) Put2(tableEntry interface{}, params *TableItemRequestBuilderPutQueryParameters) (*TableItemResponse, error) {
+func (rB *TableItemRequestBuilder) Put3(ctx context.Context, tableEntry interface{}, params *TableItemRequestBuilderPutQueryParameters) (*TableItemResponse, error) {
 	tableEntry, err := convertFromTableEntry(tableEntry)
 	if err != nil {
 		return nil, err
@@ -102,7 +64,7 @@ func (rB *TableItemRequestBuilder) Put2(tableEntry interface{}, params *TableIte
 		response: &TableItemResponse2[TableEntry]{},
 	}
 
-	err = rB.SendPut2(config.toConfiguration())
+	err = rB.SendPut3(ctx, config.toConfiguration())
 	if err != nil {
 		return nil, err
 	}